Update Materials docs (#168)
