+++ docs/src/lib/HyperElasticMaterials.md
@@ -1,0 +1,7 @@
+# AbstractHyperElasticMaterial Interface
+
+This section describes the interfaces for different hyper-elastic material types.
+
+```@autodocs
+Modules = [ONSAS.Materials.HyperElasticMaterials]
+```
