+++ docs/src/lib/LinearElasticMaterials.md
@@ -1,0 +1,7 @@
+# AbstractLinearElasticMaterial Interface
+
+This section describes the interfaces for different linear elastic material types.
+
+```@autodocs
+Modules = [ONSAS.Materials.LinearElasticMaterials]
+```
