+++ src/Materials/LinearElasticMaterials.jl
@@ -1,0 +1,46 @@
+module LinearElasticMaterials
+
+using ..Materials: AbstractMaterial
+
+export AbstractLinearElasticMaterial, lame_parameters, elasticity_modulus, shear_modulus,
+    bulk_modulus, poisson_ratio, cauchy_stress
+
+""" Abstract supertype for all elastic material models.
+
+An `AbstractLinearElasticMaterial` object facilitates the process of using elastic materials.
+
+**Common methods:**
+* [`elastic_modulus`](@ref)
+* [`poisson_ratio`](@ref)
+* [`shear_modulus`](@ref)
+* [`lamé_parameters`](@ref)
+
+**Common fields:**
+* label
+* ρ(density)
+
+"""
+abstract type AbstractLinearElasticMaterial <: AbstractMaterial end
+
+"Returns Lamé parameters `λ` and `G` from an `AbstractLinearElasticMaterial` material `m`."
+function lame_parameters(m::AbstractLinearElasticMaterial) end
+
+"Returns shear modulus G` from an `AbstractLinearElasticMaterial` material `m`."
+function shear_modulus(m::AbstractLinearElasticMaterial) end
+
+"Returns the Poisson's ratio `ν` form an `AbstractLinearElasticMaterial` material `m`."
+function poisson_ratio(m::AbstractLinearElasticMaterial) end
+
+"Returns the Elasticity modulus `E` form an `AbstractLinearElasticMaterial` material `m`."
+function elasticity_modulus(m::AbstractLinearElasticMaterial) end
+
+"Returns the bulk modulus `K` for an `AbstractLinearElasticMaterial` material `m`."
+function bulk_modulus(m::AbstractLinearElasticMaterial) end
+
+"Returns the cauchy stress tensor `σ` and the constitutive driver `∂σ∂ϵ` 
+considering a `IsotropicLinearElastic` material `m`."
+function cauchy_stress(m::AbstractLinearElasticMaterial, ϵ::AbstractMatrix) end
+
+include("IsotropicLinearElastic.jl")
+
+end
