+++ src/Materials/Materials.jl
@@ -38,67 +38,13 @@ density(m::AbstractMaterial) = m.ρ
 "Returns the `AbstractMaterial` `m` label."
 label(m::AbstractMaterial) = m.label
 
-""" Abstract supertype for all elastic material models.
+include("./LinearElasticMaterials.jl")
+@reexport using .LinearElasticMaterials
 
-An `AbstractLinearElasticMaterial` object facilitates the process of using elastic materials.
+include("./HyperElasticMaterials.jl")
+@reexport using .HyperElasticMaterials
 
-**Common methods:**
-* [`elastic_modulus`](@ref)
-* [`poisson_ratio`](@ref)
-* [`shear_modulus`](@ref)
-* [`lamé_parameters`](@ref)
 
-**Common fields:**
-* label
-* ρ(density)
-
-"""
-abstract type AbstractLinearElasticMaterial <: AbstractMaterial end
-
-"Returns Lamé parameters `λ` and `G` from an `AbstractLinearElasticMaterial` material `m`."
-function lame_parameters(m::AbstractLinearElasticMaterial) end
-
-"Returns shear modulus G` from an `AbstractLinearElasticMaterial` material `m`."
-function shear_modulus(m::AbstractLinearElasticMaterial) end
-
-"Returns the Poisson's ratio `ν` form an `AbstractLinearElasticMaterial` material `m`."
-function poisson_ratio(m::AbstractLinearElasticMaterial) end
-
-"Returns the Elasticity modulus `E` form an `AbstractLinearElasticMaterial` material `m`."
-function elasticity_modulus(m::AbstractLinearElasticMaterial) end
-
-"Returns the bulk modulus `K` for an `AbstractLinearElasticMaterial` material `m`."
-function bulk_modulus(m::AbstractLinearElasticMaterial) end
-
-include("IsotropicLinearElastic.jl")
-
-""" Abstract supertype for all hyper-elastic material models.
-
-An `AbstractHyperElasticMaterial` object facilitates the process of using hyper elastic materials.
-These materials are characterized by a strain energy function ψ  that depends only on the deformation gradient tensor `∇u`.
-
-**Common methods:**
-* [`strain_energy`](@ref)
-* [`cosserat_stress`](@ref)
-
-**Common fields:**
-* label
-* ρ(density)
-
-"""
-abstract type AbstractHyperElasticMaterial <: AbstractMaterial end
-
-"Returns the strain energy value for an `AbstractMaterial` `m`, and the Green-Lagrange 
-strain tensor `𝔼`."
-function strain_energy(m::AbstractHyperElasticMaterial, 𝔼) end
-
-"Returns the Cosserat or Second-Piola Kirchhoff stress tensor `𝕊` given an `AbstractMaterial` `m` and the 
-Green-Lagrange strain tensor `𝔼`."
-function cosserat_stress(m::AbstractMaterial, 𝔼::AbstractMatrix) end
-
-include("SVK.jl")
-include("NeoHookean.jl")
-include("HyperElastic.jl")
 
 
 end # module
